chore: js-build, js-cleanのログを改善

+++ scripts/js-clean.ts
@@ -17,7 +17,11 @@ const { values } = parseArgs({
     allowPositionals: true,
 });
 
-const rm = (path: string) => {
+const rm = async (path: string) => {
+    if (!await fs.exists(path)) {
+        console.info(`Skipping ${path} as it does not exist`);
+        return;
+    }
     fs.rm(path, { recursive: true })
         .then(() => console.info(`Successfully deleted: ${path}`))
         .catch(err => console.error(`Error cleaning ${path}: ${err.message}`))
